commands: Unbreak support for custom commands.
Did this break between python3.4 and 3.5? Or did it break when we added
an init.py? We need tests for this.

Apparently in 3.5, we cannot load "asterisklint.commands.whatever" from
another path than the already loaded "asterisklint.commands" directory.

Fixed by using importlib.util to load the file by path.